Move invertMap into ReuseAllocations
